Refactor json.hpp to use int64_t instead of int for improved code consistency

+++ include/json.hpp
@@ -272,7 +272,7 @@ namespace json
      */
     json &operator=(int l) noexcept
     {
-      value = l;
+      value = static_cast<int64_t>(l);
       return *this;
     }
